// src/api/openapi/models/HarvestedCropReportFormData.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type HarvestedCropReportFormData = {
    harvested_qty: string;
    withered_crops: string;
    date_harvested: string;
    notes?: string;
    kilogram: string;
    task_id?: string;
    images: Array<Blob>;
};